+++ src/main/java/pp/coinwash/history/domain/dto/UsageWashingDto.java
@@ -1,7 +1,7 @@
-package pp.coinwash.usage.domain.dto;
+package pp.coinwash.history.domain.dto;
 
 import lombok.Builder;
-import pp.coinwash.usage.domain.type.WashingCourse;
+import pp.coinwash.history.domain.type.WashingCourse;
 
 @Builder
 public record UsageWashingDto(
