+++ src/main/java/pp/coinwash/history/domain/repository/UsageHistoryRepository.java
@@ -1,10 +1,10 @@
-package pp.coinwash.usage.domain.repository;
+package pp.coinwash.history.domain.repository;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import pp.coinwash.usage.domain.entity.UsageHistory;
+import pp.coinwash.history.domain.entity.UsageHistory;
 
 public interface UsageHistoryRepository extends JpaRepository<UsageHistory, Long> {
 	Page<UsageHistory> findAllByCustomerId(long customerId, Pageable pageable);
