+++ src/test/java/pp/coinwash/history/service/UsageHistoryServiceTest.java
@@ -1,4 +1,4 @@
-package pp.coinwash.usage.service;
+package pp.coinwash.history.service;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.Mockito.*;
@@ -21,10 +21,10 @@ import org.springframework.data.domain.Pageable;
 import pp.coinwash.common.dto.PagedResponseDto;
 import pp.coinwash.laundry.domain.entity.Laundry;
 import pp.coinwash.machine.domain.entity.Machine;
-import pp.coinwash.usage.domain.dto.UsageHistoryRequestDto;
-import pp.coinwash.usage.domain.dto.UsageHistoryResponseDto;
-import pp.coinwash.usage.domain.entity.UsageHistory;
-import pp.coinwash.usage.domain.repository.UsageHistoryRepository;
+import pp.coinwash.history.domain.dto.UsageHistoryRequestDto;
+import pp.coinwash.history.domain.dto.UsageHistoryResponseDto;
+import pp.coinwash.history.domain.entity.UsageHistory;
+import pp.coinwash.history.domain.repository.UsageHistoryRepository;
 
 @ExtendWith(MockitoExtension.class)
 class UsageHistoryServiceTest {
